requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
dnspython>=2.4.0
python-whois>=0.8.0
aiohttp>=3.9.0
tldextract>=5.1.0
colorama>=0.4.6
rich>=13.7.0
PyMuPDF>=1.23.0
python-docx>=1.1.0
openpyxl>=3.1.0
pdfminer.six>=20221105
chardet>=5.2.0
fake-useragent>=1.4.0
urllib3>=2.1.0
